requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
sniffio==1.3.0
SQLAlchemy==1.4.43
SQLAlchemy-Utils==0.38.3
starlette==0.20.4
typing_extensions==4.4.0
uvicorn==0.19.0
anyio==3.6.2
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
SQLAlchemy-Utils==0.38.3
stack-data==0.6.1
starlette==0.20.4
toml==0.10.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wcwidth==0.2.5
anyio==3.6.2
asttokens==2.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
decorator==5.1.1
executing==1.2.0
fastapi==0.86.0
greenlet==2.0.1
h11==0.14.0
idna==3.4
ipdb==0.13.9
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.32
psycopg2-binary==2.9.5
pure-eval==0.2.2
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.43
SQLAlchemy-Utils==0.38.3
stack-data==0.6.1
starlette==0.20.4
toml==0.10.2
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
wcwidth==0.2.5
boto3==1.24.59
aws-secretsmanager-caching==1.1.1.5
email-validator==1.3.0
pytz==2022.6
Werkzeug==2.2.2
PyJWT==2.6.0